+++ meetup2xibo/updater/injector.py
@@ -3,7 +3,7 @@
 from .logging_context import LoggingContext
 from .meetup2xibo import Meetup2Xibo, XiboSessionProcessor, \
         XiboEventCrudProcessor
-from .meetup_api import MeetupEventsRetriever, meetup_iso_offset_time
+from .meetup_api import MeetupEventsRetriever
 from .location_builder import LocationBuilder
 from .location_chooser import LocationChooser
 from .event_converter import EventConverter
@@ -19,7 +19,7 @@ from .xibo_dataset_id_finder import XiboDatasetIdFinder
 from .xibo_event import XiboEvent, XiboEventColumnNameManager, \
         XiboEventColumnIdManager
 from .xibo_event_crud import XiboEventCrud
-from .anti_flapper import AntiFlapper, iso_offset_time
+from .anti_flapper import AntiFlapper
 from ahocorasick import Automaton
 from requests_toolbelt import user_agent
 from datetime import datetime
@@ -99,7 +99,8 @@ def inject_automaton():
 def inject_event_converter(application_scope):
     """Return an event converter configured by an application scope."""
     return EventConverter(
-        inject_location_chooser(application_scope))
+        inject_location_chooser(application_scope),
+        inject_date_time_creator(application_scope))
 
 
 def inject_xibo_api_url_builder(application_scope):
@@ -274,14 +275,14 @@ def inject_special_location_monitor(application_scope):
     return SpecialLocationMonitor(application_scope.special_locations_dict)
 
 
-def inject_tzinfo():
+def inject_tzinfo(application_scope):
     """Return timezone info configured by an application scope."""
     return timezone(application_scope.timezone)
 
 
 def inject_date_time_creator(application_scope):
     """Return a date/time creator configured by an application scope."""
-    return DateTimeCreator(inject_tzinfo())
+    return DateTimeCreator(inject_tzinfo(application_scope))
 
 
 def inject_recent_limit(application_scope):